+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [2.5.6] - 2018-04-18
+## [2.5.6] - 2018-04-20
 
-### Added
-- added member service and MemberUtil::findActiveByGroups
+### Fixed
+- `setContext` and `getContext` handling in `AbstractChoice`
+
+### Changed
+- `TwigTemplateChoice` now also collects twig templates inside root `templates/` directory and returns the path to the template e.g. `templates/news_full.html.twig`
 
 ## [2.5.5] - 2018-04-17